backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
celery>=5.3
redis>=5.0 
python-dotenv
drf-spectacular
requests